requirements.txt
asgiref==3.3.4
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
psycopg2
djangorestframework
dj_rest_auth
coreapi
django_stubs
mypy
requests
django-rest-knox==4.1.0